清理部分代码告警

+++ services/audio_service/common/src/linear_pos_time_model.cpp
@@ -125,7 +125,7 @@ int64_t LinearPosTimeModel::GetTimeOfPos(uint64_t posInFrame)
             AUDIO_WARNING_LOG("posInFrame %{public}" PRIu64" is too"
                 " small, stampFrame: %{public}" PRIu64"", posInFrame, stampFrame_);
         }
-        deltaFrame = stampFrame_ - posInFrame;
+        deltaFrame = static_cast<int64_t>(stampFrame_ - posInFrame);
         return stampNanoTime_ - deltaFrame * NANO_COUNT_PER_SECOND / (int64_t)sampleRate_;
     }
     return invalidTime;
